Move the rest of the base lib directory to lib.generator to avoid further issues
