+++ java-generator/src/main/scala/models/generator/JavaUtil.scala
@@ -1,6 +1,6 @@
 package models.generator
 
-import lib.Text
+import lib.generator.Text
 
 /**
  *
