+++ lib/src/main/scala/generator/Constants.scala
@@ -1,4 +1,4 @@
-package lib
+package lib.generator
 
 object Constants {
