+++ lib/src/main/scala/generator/Methods.scala
@@ -1,4 +1,4 @@
-package lib
+package lib.generator
 
 object Methods {
