+++ lib/src/main/scala/generator/VersionTag.scala
@@ -1,4 +1,4 @@
-package lib
+package lib.generator
 
 object VersionTag {
