+++ scala-generator/src/main/scala/models/Headers.scala
@@ -2,8 +2,8 @@ package scala.models
 
 import scala.generator.ScalaUtil
 import com.bryzek.apidoc.generator.v0.models.InvocationForm
-import lib.VersionTag
-import lib.Text._
+import lib.generator.VersionTag
+import lib.generator.Text._
 
 case class Headers(
   form: InvocationForm
