+++ scala-generator/src/main/scala/models/Play2Json.scala
@@ -1,6 +1,6 @@
 package scala.models
 
-import lib.Text._
+import lib.generator.Text._
 import scala.generator.{PrimitiveWrapper, ScalaDatatype, ScalaModel, ScalaPrimitive, ScalaService, ScalaUnion, ScalaUnionType}
 
 case class Play2Json(
@@ -16,7 +16,7 @@ case class Play2Json(
       ssd.models.map(readersAndWriters(_)).mkString("\n\n"),
       PrimitiveWrapper(ssd).wrappers.map(w => readersAndWriters(w.model)).mkString("\n\n"),
       ssd.unions.map(readersAndWriters(_)).mkString("\n\n")
-    ).filter(!_.trim.isEmpty).mkString("\n\n")    
+    ).filter(!_.trim.isEmpty).mkString("\n\n")
   }
 
   private def readersAndWriters(union: ScalaUnion): String = {
