+++ scala-generator/src/main/scala/models/generator/ScalaCaseClasses.scala
@@ -3,9 +3,8 @@ package scala.generator
 import scala.models.ApidocComments
 import com.bryzek.apidoc.generator.v0.models.{File, InvocationForm}
 import com.bryzek.apidoc.spec.v0.models.Service
-import lib.Text._
-import lib.generator.CodeGenerator
-import generator.ServiceFileNames
+import lib.generator.Text._
+import lib.generator.{CodeGenerator, ServiceFileNames}
 
 object ScalaCaseClasses extends CodeGenerator {
