+++ scala-generator/src/main/scala/models/generator/ScalaClientObject.scala
@@ -1,6 +1,6 @@
 package scala.generator
 
-import lib.Text._
+import lib.generator.Text._
 
 object ScalaClientObject {
