+++ scala-generator/src/main/scala/models/generator/ScalaGeneratorUtil.scala
@@ -2,7 +2,7 @@ package scala.generator
 
 import com.bryzek.apidoc.spec.v0.models.ParameterLocation
 import lib.generator.Datatype
-import lib.Text._
+import lib.generator.Text._
 
 case class ScalaGeneratorUtil(config: ScalaClientMethodConfig) {
